Changes at the App after first big test.
After testing the software system at the middle-German kanoe
championship, many improvements were made. An app-crashing Bug was
fixed. The GUI color was changed to make it easier to read. The
possibility to start or stopp the next or last starter as well as the
current was added to make it easier to start or stopp a close group.
Vertical spaces were added. Also, a software-based "back" button was
added. Finally, minor problems were fixed.